test: Split the upgrade tests across the units they cover
Move installation classification, option parsing, the update prompt and
runDartInstall into their own suites.
