+++ tools/serverpod_cli/test/commands/upgrade/cli_installation_test.dart
@@ -39,6 +39,36 @@ String _createExecutable(final Directory directory) {
   return path;
 }
 
+/// Resolves a path the way [CliInstallation.resolve] does before classifying.
+String _canonicalize(final String path) {
+  try {
+    return File(path).resolveSymbolicLinksSync();
+  } catch (_) {
+    return p.canonicalize(path);
+  }
+}
+
+/// Builds the tree `dart install` leaves behind, and returns the bundle
+/// executable it installed.
+String _createAppBundle(final Directory root) {
+  final bundleBin = Directory(
+    p.join(
+      root.path,
+      'install',
+      'app-bundles',
+      'serverpod_cli',
+      'hosted',
+      '4.0.0-rc.1',
+      'bundle',
+      'bin',
+    ),
+  )..createSync(recursive: true);
+
+  final executable = p.join(bundleBin.path, 'serverpod');
+  File(executable).writeAsStringSync('');
+  return executable;
+}
+
 void main() {
   group('Given several directories on PATH containing the executable,', () {
     test(
@@ -184,6 +214,111 @@ void main() {
             p.join(p.separator, 'pub-cache', 'bin', 'serverpod'),
           ),
           isFalse,
+        );
+      },
+    );
+  });
+
+  group('Given an executable dart install put on PATH,', () {
+    late String bundleExecutable;
+    late Directory binDirectory;
+    late String installRoot;
+
+    setUp(() {
+      final root = _createTempDirectory();
+      bundleExecutable = _createAppBundle(root);
+      installRoot = _canonicalize(p.join(root.path, 'install'));
+      binDirectory = Directory(p.join(root.path, 'install', 'bin'))
+        ..createSync(recursive: true);
+    });
+
+    test(
+      'when the bin entry is a symlink, as on macOS and Linux, '
+      'then an upgrade replaces the running executable',
+      () {
+        final managed = p.join(binDirectory.path, 'serverpod');
+        Link(managed).createSync(bundleExecutable);
+
+        expect(
+          classifyInstallation(
+            runningExecutable: _canonicalize(bundleExecutable),
+            runningScript: '',
+            installRoot: installRoot,
+            globalPackagesRoot: null,
+          ),
+          CliInstallationKind.managed,
+        );
+      },
+    );
+
+    test(
+      'when the bin entry is a .bat wrapper, as on Windows, '
+      'then an upgrade replaces the running executable',
+      () {
+        // emulate pkg/dartdev/lib/src/install/file_system.dart (not symlink)
+        final managed = p.join(binDirectory.path, 'serverpod.bat');
+        File(managed).writeAsStringSync(
+          '@ECHO OFF\n'
+          'REM target_file_path_marker\n'
+          '"$bundleExecutable" %*\n'
+          'EXIT /B %ERRORLEVEL%\n',
+        );
+
+        expect(
+          classifyInstallation(
+            runningExecutable: _canonicalize(bundleExecutable),
+            runningScript: '',
+            installRoot: installRoot,
+            globalPackagesRoot: null,
+          ),
+          CliInstallationKind.managed,
+        );
+      },
+    );
+  });
+
+  group('Given the CLI was started by a pub global binstub,', () {
+    test(
+      'when classifying the installation, '
+      'then it is not reported as a source checkout',
+      () {
+        final root = _createTempDirectory();
+
+        final sdkBin = Directory(p.join(root.path, 'dart-sdk', 'bin'))
+          ..createSync(recursive: true);
+        final dartExecutable = p.join(sdkBin.path, 'dart');
+        File(dartExecutable).writeAsStringSync('');
+
+        final pubCacheBin = Directory(p.join(root.path, '.pub-cache', 'bin'))
+          ..createSync(recursive: true);
+        File(p.join(pubCacheBin.path, 'serverpod')).writeAsStringSync(
+          '#!/usr/bin/env sh\n'
+          'dart pub global run serverpod_cli:serverpod_cli "\$@"\n',
+        );
+
+        final globalPackages = Directory(
+          p.join(root.path, '.pub-cache', 'global_packages'),
+        )..createSync(recursive: true);
+        final snapshotDirectory = Directory(
+          p.join(globalPackages.path, 'serverpod_cli', 'bin'),
+        )..createSync(recursive: true);
+        final snapshot = p.join(
+          snapshotDirectory.path,
+          'serverpod_cli.dart-3.13.2.snapshot',
+        );
+        File(snapshot).writeAsStringSync('');
+
+        expect(
+          classifyInstallation(
+            runningExecutable: _canonicalize(dartExecutable),
+            runningScript: _canonicalize(snapshot),
+            installRoot: _canonicalize(p.join(root.path, 'install')),
+            globalPackagesRoot: _canonicalize(globalPackages.path),
+          ),
+          isNot(CliInstallationKind.source),
+          reason:
+              'These users need the deactivate hint, which only the '
+              'foreign branch prints.',
         );
       },
     );
