+++ tools/serverpod_cli/test/commands/upgrade/upgrade_command_test.dart
@@ -2,6 +2,7 @@ import 'package:pub_semver/pub_semver.dart';
 import 'package:serverpod_cli/src/commands/upgrade.dart';
 import 'package:serverpod_cli/src/commands/upgrade/upgrade_target.dart';
 import 'package:test/test.dart';
+import 'package:yaml/yaml.dart';
 
 void main() {
   group('Given an UpgradeCommand,', () {
@@ -58,5 +59,39 @@ void main() {
         expect(config.value(UpgradeOption.force), isTrue);
       },
     );
+  });
+
+  group('Given a value passed to --version,', () {
+    late UpgradeCommand command;
+
+    setUp(() => command = UpgradeCommand());
+
+    Iterable<Object?> errorsFor(final String value) => command
+        .resolveConfiguration(command.argParser.parse(['--version', value]))
+        .errors;
+
+    for (final value in const [
+      '4.0.0',
+      '4.0.0-beta.4',
+      '4.0.0+1',
+      '1.2.3-rc.1+build.5',
+    ]) {
+      test('when "$value" is given, then dart install can parse it too', () {
+        expect(errorsFor(value), isEmpty);
+        expect(() => loadYaml(value), returnsNormally);
+      });
+    }
+
+    for (final value in const [
+      '>=4.0.0 <5.0.0',
+      '>=4.0.0',
+      '>4.0.0',
+      '^4.0.0',
+      'any',
+    ]) {
+      test('when the constraint "$value" is given, then it is rejected', () {
+        expect(errorsFor(value), isNotEmpty);
+      });
+    }
   });
 }
